requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
idna==2.7
multidict==4.4.2
requests==2.19.1
tinydb==3.11.1
urllib3==1.23
yarl==1.2.6
html5lib==1.0.1
coloredlogs==10.0
